Change test classes names
