+++ dartagnan/src/test/java/com/dat3m/dartagnan/benchmarking/assume/ARM8.java
@@ -2,10 +2,7 @@ package com.dat3m.dartagnan.benchmarking.assume;
 
 import com.dat3m.dartagnan.utils.Result;
 import com.dat3m.dartagnan.utils.rules.CSVLogger;
-import com.dat3m.dartagnan.utils.rules.Provider;
-import com.dat3m.dartagnan.utils.rules.Providers;
 import com.dat3m.dartagnan.verification.solving.AssumeSolver;
-import com.dat3m.dartagnan.wmm.Wmm;
 import com.dat3m.dartagnan.benchmarking.AbstractDartagnanTest;
 import com.dat3m.dartagnan.configuration.Arch;
 import org.junit.Test;
@@ -20,32 +17,27 @@ import static com.dat3m.dartagnan.configuration.Arch.*;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
-public class DartagnanIMM extends AbstractDartagnanTest {
+public class ARM8 extends AbstractDartagnanTest {
 
-    public DartagnanIMM(String name, Arch target, Result expected) {
+    public ARM8(String name, Arch target, Result expected) {
         super(name, target, expected);
-    }
-
-    @Override
-    protected Provider<Wmm> getWmmProvider() {
-        return Providers.createWmmFromName(() -> "imm");
     }
 
 	@Parameterized.Parameters(name = "{index}: {0}, target={1}")
     public static Iterable<Object[]> data() throws IOException {
 		return Arrays.asList(new Object[][]{
-	            {"locks/ttas-5", IMM, UNKNOWN},
-	            {"locks/ticketlock-6", IMM, PASS},
-                {"locks/mutex-4", IMM, UNKNOWN},
-                {"locks/spinlock-5", IMM, UNKNOWN},
-                {"locks/linuxrwlock-3", IMM, UNKNOWN},
-                {"locks/mutex_musl-4", IMM, UNKNOWN},
-                {"lfds/safe_stack-3", IMM, FAIL},
-                {"lfds/chase-lev-5", IMM, PASS},
-                {"lfds/dglm-3", IMM, UNKNOWN},
-                {"lfds/harris-3", IMM, UNKNOWN},
-                {"lfds/ms-3", IMM, UNKNOWN},
-                {"lfds/treiber-3", IMM, UNKNOWN},
+	            {"locks/ttas-5", ARM8, UNKNOWN},
+	            {"locks/ticketlock-6", ARM8, PASS},
+                {"locks/mutex-4", ARM8, UNKNOWN},
+                {"locks/spinlock-5", ARM8, UNKNOWN},
+                {"locks/linuxrwlock-3", ARM8, UNKNOWN},
+                {"locks/mutex_musl-4", ARM8, UNKNOWN},
+                {"lfds/safe_stack-3", ARM8, FAIL},
+                {"lfds/chase-lev-5", ARM8, PASS},
+                {"lfds/dglm-3", ARM8, UNKNOWN},
+                {"lfds/harris-3", ARM8, UNKNOWN},
+                {"lfds/ms-3", ARM8, UNKNOWN},
+                {"lfds/treiber-3", ARM8, UNKNOWN},
 		});
     }
