+++ dartagnan/src/test/java/com/dat3m/dartagnan/benchmarking/assume/Power.java
@@ -17,27 +17,27 @@ import static com.dat3m.dartagnan.configuration.Arch.*;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
-public class DartagnanARM8 extends AbstractDartagnanTest {
+public class Power extends AbstractDartagnanTest {
 
-    public DartagnanARM8(String name, Arch target, Result expected) {
+    public Power(String name, Arch target, Result expected) {
         super(name, target, expected);
     }
 
 	@Parameterized.Parameters(name = "{index}: {0}, target={1}")
     public static Iterable<Object[]> data() throws IOException {
 		return Arrays.asList(new Object[][]{
-	            {"locks/ttas-5", ARM8, UNKNOWN},
-	            {"locks/ticketlock-6", ARM8, PASS},
-                {"locks/mutex-4", ARM8, UNKNOWN},
-                {"locks/spinlock-5", ARM8, UNKNOWN},
-                {"locks/linuxrwlock-3", ARM8, UNKNOWN},
-                {"locks/mutex_musl-4", ARM8, UNKNOWN},
-                {"lfds/safe_stack-3", ARM8, FAIL},
-                {"lfds/chase-lev-5", ARM8, PASS},
-                {"lfds/dglm-3", ARM8, UNKNOWN},
-                {"lfds/harris-3", ARM8, UNKNOWN},
-                {"lfds/ms-3", ARM8, UNKNOWN},
-                {"lfds/treiber-3", ARM8, UNKNOWN},
+            {"locks/ttas-5", POWER, UNKNOWN},
+            {"locks/ticketlock-6", POWER, PASS},
+            {"locks/mutex-4", POWER, UNKNOWN},
+            {"locks/spinlock-5", POWER, UNKNOWN},
+            {"locks/linuxrwlock-3", POWER, UNKNOWN},
+            {"locks/mutex_musl-4", POWER, UNKNOWN},
+            {"lfds/safe_stack-3", POWER, FAIL},
+            {"lfds/chase-lev-5", POWER, PASS},
+            {"lfds/dglm-3", POWER, UNKNOWN},
+            {"lfds/harris-3", POWER, UNKNOWN},
+            {"lfds/ms-3", POWER, UNKNOWN},
+            {"lfds/treiber-3", POWER, UNKNOWN},
 		});
     }
